test: add buildTerraformOptions, generateUniqueID and createTestFolder
